Taken Liquidity Bug
So in MT4, there was a bug where Taken Liquidity would not spawn due to an Object with the same being present. So to fix this I basically created a var in that is TL (Taken Liquidity) and that would continue to grow and count - 1 the current amount of taken Liquidity and count would be the amount of taken Liquidity +1 which is the Liquidity soon to be taken. I Also did it in MT5 because maybe it will experience the same thing.